+++ tests/unit/small_text/integrations/pytorch/classifiers/test_kimcnn.py
@@ -24,12 +24,10 @@ except PytorchNotFoundError:
 @pytest.mark.pytorch
 class KimCNNTest(unittest.TestCase):
 
-    def _get_clf(self, num_classes=2, early_stopping=5, early_stopping_acc=-1):
+    def _get_clf(self, num_classes=2):
         embedding_matrix = torch.rand(10, 20)
         return KimCNNClassifier(num_classes, embedding_matrix=embedding_matrix, num_epochs=2,
-                                out_channels=15, max_seq_len=20, kernel_heights=[2, 3],
-                                device='cpu', early_stopping=early_stopping,
-                                early_stopping_acc=early_stopping_acc)
+                                out_channels=15, max_seq_len=20, kernel_heights=[2, 3], device='cpu')
 
     def test_init_default_parameters(self):
         num_classes = 2
@@ -53,8 +51,6 @@ class KimCNNTest(unittest.TestCase):
         np.testing.assert_equal(embedding_matrix, classifier.embedding_matrix)
         self.assertEqual(0, classifier.padding_idx)
         self.assertEqual([3, 4, 5], classifier.kernel_heights)
-        self.assertEqual(5, classifier.early_stopping)
-        self.assertEqual(-1, classifier.early_stopping_acc)
         self.assertIsNone(classifier.class_weight)
         self.assertFalse(classifier.compile_model)
         self.assertIsNone(classifier.model)
@@ -72,8 +68,6 @@ class KimCNNTest(unittest.TestCase):
         validation_set_size = 0.15
         embedding_matrix = np.random.rand(5, 10)
         padding_idx = 1
-        early_stopping = 10
-        early_stopping_acc = 0.95
         class_weight = 'balanced'
         compile_model = False
 
@@ -83,8 +77,6 @@ class KimCNNTest(unittest.TestCase):
                                       filter_padding=filter_padding, dropout=dropout,
                                       validation_set_size=validation_set_size,
                                       embedding_matrix=embedding_matrix, padding_idx=padding_idx,
-                                      early_stopping=early_stopping,
-                                      early_stopping_acc=early_stopping_acc,
                                       class_weight=class_weight,
                                       compile_model=compile_model)
 
@@ -103,8 +95,6 @@ class KimCNNTest(unittest.TestCase):
         self.assertEqual(validation_set_size, classifier.validation_set_size)
         np.testing.assert_equal(embedding_matrix, classifier.embedding_matrix)
         self.assertEqual(padding_idx, classifier.padding_idx)
-        self.assertEqual(early_stopping, classifier.early_stopping)
-        self.assertEqual(early_stopping_acc, classifier.early_stopping_acc)
         self.assertEqual(class_weight, classifier.class_weight)
         self.assertFalse(classifier.compile_model)
         self.assertIsNone(classifier.model)
@@ -119,16 +109,13 @@ class KimCNNTest(unittest.TestCase):
         validation_set_size = 0.15
         embedding_matrix = None
         padding_idx = 1
-        early_stopping = 10
-        early_stopping_acc = 0.95
 
         with self.assertRaises(ValueError):
             KimCNNClassifier(num_classes, embedding_matrix=embedding_matrix, device='cpu',
                              num_epochs=num_epochs, mini_batch_size=mini_batch_size,
                              max_seq_len=max_seq_len, out_channels=out_channels,
                              dropout=dropout, validation_set_size=validation_set_size,
-                             padding_idx=padding_idx, early_stopping=early_stopping,
-                             early_stopping_acc=early_stopping_acc)
+                             padding_idx=padding_idx)
 
     def test_fit_without_validation_set(self):
         dataset = random_text_classification_dataset(10)
@@ -213,59 +200,3 @@ class KimCNNTest(unittest.TestCase):
 
         with self.assertRaisesRegex(ValueError, 'Weights must be greater zero.'):
             classifier.fit(train_set, validation_set=validation_set, weights=weights)
-
-    # TODO: remove this in 2.0.0
-    def test_fit_with_early_stopping_fallback_default_kwargs(self):
-        train = random_text_classification_dataset(8)
-        valid = random_text_classification_dataset(2)
-
-        classifier = self._get_clf()
-        with patch.object(classifier, '_fit_main') as fit_main_mock:
-            classifier.fit(train, validation_set=valid)
-            fit_main_mock.assert_called()
-            self.assertIsNone(classifier.class_weights_)
-
-            call_args = fit_main_mock.call_args[0]
-            self.assertTrue(isinstance(call_args[3], EarlyStopping))
-            self.assertEqual('val_loss', call_args[3].metric.name)
-            self.assertEqual(5, call_args[3].patience)
-
-    # TODO: remove this in 2.0.0
-    def test_fit_with_early_stopping_fallback_deprecated_kwargs(self):
-        train = random_text_classification_dataset(8)
-        valid = random_text_classification_dataset(2)
-
-        early_stopping_no_improvement = 8
-        early_stopping_acc = 0.98
-
-        classifier = self._get_clf(early_stopping=early_stopping_no_improvement,
-                                   early_stopping_acc=early_stopping_acc)
-        with patch.object(classifier, '_fit_main') as fit_main_mock:
-            classifier.fit(train, validation_set=valid)
-            fit_main_mock.assert_called()
-            self.assertIsNone(classifier.class_weights_)
-
-            call_args = fit_main_mock.call_args[0]
-            self.assertTrue(isinstance(call_args[3], EarlyStoppingOrCondition))
-
-            first_handler = call_args[3].early_stopping_handlers[0]
-            self.assertEqual('val_loss', first_handler.metric.name)
-            self.assertEqual(early_stopping_no_improvement, first_handler.patience)
-
-            second_handler = call_args[3].early_stopping_handlers[1]
-            self.assertEqual('train_acc', second_handler.metric.name)
-            self.assertEqual(-1, second_handler.patience)
-
-    # TODO: remove this in 2.0.0
-    def test_fit_with_early_stopping_and_fall_back_simultaneously(self):
-        dataset = random_text_classification_dataset(10)
-
-        early_stopping = EarlyStopping(Metric('val_loss'))
-
-        classifier = self._get_clf(early_stopping=7)
-        with self.assertWarnsRegex(UserWarning, r'Both the fit\(\) argument early_stopping'):
-            classifier.fit(dataset, early_stopping=early_stopping)
-
-        classifier = self._get_clf(early_stopping_acc=0.98)
-        with self.assertWarnsRegex(UserWarning, r'Both the fit\(\) argument early_stopping'):
-            classifier.fit(dataset, early_stopping=early_stopping)